requirements.txt
fastapi
uvicorn
pandas
gspread
google-auth
google-api-python-client
